requirements.txt
streamlit
PyMuPDF
pandas
google-generativeai

